fix: rename exam and sample model files
