+++ app/Models/Sample.php
